CDPT-1193 external user access (#942)
* Create external user table.

* Add initial external user model.

* Add unique email constraint.

* Session creator checks and returns correct user type.

* Save login count and timestamp only with MoJ users.

* Login - redirect to profile if internal user, to homepage if external user.

* Login - get session key for person type.

* Set external users with read-only access.

* Add external user tests for:
- Group editing feature
- Token model validation
- Token logon service
- Login feature
- External user model
- Versions
- Home page
- Person maintenance
- Suggestions

+++ app/controllers/application_controller.rb
@@ -68,7 +68,11 @@ private
   end
 
   def desired_path(person)
-    session.delete(:desired_path) || person_path(person, prompt: :profile)
+    if person.is_a?(Person)
+      session.delete(:desired_path) || person_path(person, prompt: :profile)
+    else
+      home_path
+    end
   end
 
   def i18n_flash(type, *partial_key, **options)
